+++ src/utils/toDiscordMarkdown.ts
@@ -1,11 +1,27 @@
+function processLinks(text: string) {
+  let result = text.replace(/\<(?<linkText>.+?@.+?\..+?)\>/g, '$<linkText>')
+
+  for (const match of text.matchAll(
+    /\[(?<linkText>.+?)\]: (?<linkTarget>http.+?)\n/g
+  )) {
+    const linkText = match.groups?.['linkText']
+    const linkTarget = match.groups?.['linkTarget']
+
+    result = result.replace(
+      new RegExp(`(?<!\`)\\[(?<linkText>${linkText})\\](?!\\()`, 'g'),
+      `[$<linkText>](${linkTarget})`
+    )
+  }
+
+  // Remove square brackets from non-link text (i.e. missing links)
+  result = result.replace(/(?<!\`)\[(?<linkText>.+?)\](?!\()/g, '$<linkText>')
+
+  return result
+}
+
 export default function toDiscordMarkdown(text: string) {
   return (
-    text
-      // Remove links that are not supported in Discord Markdown.
-      .replace(/\[.+?]: http.+\n/g, '')
-      .replace(/\[(?<linkText>.+?)](?!\()/g, '$<linkText>')
-      // `mailto:` links are not supported in Discord Markdown.
-      .replace(/\<(?<linkText>.+?@.+?\..+?)\>/g, '$<linkText>')
+    processLinks(text)
       // Horizontal rules
       .replace(/\n\n-{3,}/g, '\n\n** **\n' + ' '.repeat(50))
       // Unwrap wrapped lines.